requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
astroid==2.3.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.1
blinker==1.4
Brotli==1.0.9
cachelib==0.2.0
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2019.9.11
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
cryptography==3.2.1
cssselect2==0.4.1
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
docusign-click==1.0.0
docusign-esign==3.10.0rc1
docusign-monitor==1.0.0b0
docusign-rooms==1.1.0
entrypoints==0.3
fastapi==0.67.0
Fiona==1.8.20
Flask==1.1.1
Flask-OAuthlib==0.9.6
Flask-WTF==0.14.3
fonttools==4.25.2
geopandas==0.9.0
gitdb==4.0.7
GitPython==3.1.18
h11==0.12.0
html2image==2.0.1
html5lib==1.1
idna==2.8
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
munch==2.5.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.4.0
numpy==1.21.1
oauthlib==2.1.0
opyrator==0.0.12
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pdfkit==0.6.1
pdfminer.six==20200517
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2018.11.26
plotly==5.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py-oauth2==0.0.10
pyarrow==5.0.0
pycparser==2.19
pycryptodome==3.10.1
pydantic==1.8.2
pydeck==0.6.2
pydyf==0.0.3
Pygments==2.9.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
pyproj==3.1.0
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.1.0
Send2Trash==1.7.1
Shapely==1.7.1
six==1.13.0
smmap==4.0.0
sortedcontainers==2.4.0
starlette==0.14.2
streamlit==0.86.0
tenacity==8.0.1
terminado==0.10.1
testpath==0.5.0
tinycss2==1.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
uvicorn==0.14.0
validators==0.18.2
virtualenv==16.7.5
virtualenv-clone==0.5.3
Wand==0.6.6
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.3.3
zopfli==0.1.8
